+++ src/main/MQ2DeveloperTools.cpp
@@ -1920,6 +1920,28 @@ public:
 				ImGui::TreePop();
 			}
 
+#if HAS_ALTERNATE_PERSONAS
+			ImGui::TableNextRow();
+			ImGui::TableNextColumn();
+
+			if (ImGui::TreeNode("Personas"))
+			{
+				for (int i = 0; i < MAX_PLAYER_CLASSES; ++i)
+				{
+					ImGui::TableNextRow();
+					ImGui::TableNextColumn();
+
+					ImGui::Text("%s", GetClassDesc(i + 1));
+
+					ImGui::TableNextColumn();
+					ImGui::Text("%d", pLocalPC->ProfileManager.GetAltClassLevel(i));
+					
+				}
+
+				ImGui::TreePop();
+			}
+#endif
+
 			ImGui::EndTable();
 		}
 	}
@@ -2688,6 +2710,24 @@ public:
 				ImGui::TableNextRow();
 				ImGui::TableNextColumn(); ImGui::Text("Heroic 100 Slots");
 				ImGui::TableNextColumn(); ImGui::Text("%d", eq.Heroic100Slots);
+
+				ImGui::TableNextRow();
+				ImGui::TableNextColumn(); ImGui::Text("Legacy Characters Ruleset");
+				ImGui::TableNextColumn(); ImGui::Text("%d", eq.LegacyCharactersRuleset);
+
+				ImGui::TableNextRow();
+				ImGui::TableNextColumn(); ImGui::Text("Num Max Characters");
+				ImGui::TableNextColumn(); ImGui::Text("%d", eq.NumMaxCharacters);
+
+				ImGui::TableNextRow();
+				ImGui::TableNextColumn(); ImGui::Text("Legacy Experience Bonus");
+				ImGui::TableNextColumn(); ImGui::Text("%d", eq.LegacyExperienceBonus);
+
+#if HAS_ALTERNATE_PERSONAS
+				ImGui::TableNextRow();
+				ImGui::TableNextColumn(); ImGui::Text("Num Available Personas");
+				ImGui::TableNextColumn(); ImGui::Text("%d", eq.NumAvailablePersonas);
+#endif
 
 				ImGui::TreePop();
 			}