+++ src/main/datatypes/MQ2CharacterType.cpp
@@ -353,6 +353,7 @@ enum class CharacterMembers
 	RaidLeaderExp,
 	RaidLeaderPoints,
 	PctRaidLeaderExp,
+	PersonaLevel,
 };
 
 enum class CharacterMethods
@@ -696,6 +697,7 @@ MQ2CharacterType::MQ2CharacterType() : MQ2Type("character")
 	ScopedTypeMember(CharacterMembers, RaidLeaderExp);
 	ScopedTypeMember(CharacterMembers, RaidLeaderPoints);
 	ScopedTypeMember(CharacterMembers, PctRaidLeaderExp);
+	ScopedTypeMember(CharacterMembers, PersonaLevel);
 
 	ScopedTypeMethod(CharacterMethods, Stand);
 	ScopedTypeMethod(CharacterMethods, Sit);
@@ -4124,6 +4126,38 @@ bool MQ2CharacterType::GetMember(MQVarPtr VarPtr, const char* Member, char* Inde
 #endif
 		Dest.Type = pFloatType;
 		return true;
+
+	case CharacterMembers::PersonaLevel: {
+		Dest.Type = pIntType;
+		Dest.Int = 0;
+#if HAS_ALTERNATE_PERSONAS
+		if (Index[0])
+		{
+			int classId = GetIntFromString(Index, -1);
+
+			if (classId != -1)
+			{
+				if (classId >= eqlib::Warrior && classId <= eqlib::Berserker)
+				{
+					Dest.Int = pLocalPC->GetPersonaLevel(classId);
+					return true;
+				}
+			}
+			else
+			{
+				for (int i = eqlib::Warrior; i < eqlib::Berserker; ++i)
+				{
+					if (ci_equals(Index, ClassInfo[i].ShortName))
+					{
+						Dest.Int = pLocalPC->GetPersonaLevel(i);
+						return true;
+					}
+				}
+			}
+		}
+#endif
+		return true;
+	}
 
 	default:
 		return false;